separate interpreter and constrainer
